backend/requirements.txt
lunar-python
colorama
fastapi
uvicorn
pydantic
sqlalchemy
alembic
python-dotenv
httpx
openai
